Изменение типов полей и изменение связей, добавление методов в репо,сервис и контроллер

+++ virtualBookStore/src/main/java/com/example/virtualBookStore/DTO/bookDto/BookDto.java
@@ -8,5 +8,5 @@ public record BookDto (
         String author,
         BigDecimal price,
         String description,
-        double rating
+        BigDecimal rating
 ){ }
